Canvas set to 450x450. Classes separated into files.
