+++ site/index.html
@@ -12,9 +12,9 @@
         <div>Total Score: <span id="totalScore">0</span></div>
     </div>
     <div id="container">
-        <canvas id="puzzleCanvas" width="400" height="400"></canvas>
+        <canvas id="puzzleCanvas" width="450" height="450"></canvas>
     </div>
     <button id="solveButton">Solve Puzzle</button>
-    <script src="./src/game.js"></script>
+    <script type="module" src="./src/main.js"></script>
 </body>
 </html>
