+++ site/src/ClassicPuzzle.js
@@ -1,0 +1,25 @@
+import { Puzzle } from './Puzzle.js';
+
+// Klasse für klassische Puzzles
+export class ClassicPuzzle extends Puzzle {
+    // Implementierung der spezifischen Logik für klassische Puzzles
+    init() {
+        // Initialisierung des klassischen Puzzles
+    }
+
+    draw() {
+        // Zeichnen des klassischen Puzzles
+    }
+
+    handleUserInput(event) {
+        // Behandlung der Benutzereingaben für das klassische Puzzle
+    }
+
+    checkCompletion() {
+        // Überprüfen, ob das klassische Puzzle gelöst ist
+    }
+
+    solvePuzzle() {
+        // Lösen des klassischen Puzzles
+    }
+}
